+++ src/App.tsx
@@ -1,13 +1,10 @@
 import { RecoilRoot } from "recoil";
-import { SearchParkProvider } from "./providers/SearchParkProvider";
 import { Router } from "./router/Router";
 
 export const App = () => {
   return (
     <RecoilRoot>
-      <SearchParkProvider>
-        <Router />
-      </SearchParkProvider>
+      <Router />
     </RecoilRoot>
   );
 };
